refactor: restore pngjs codec
