+++ src/utils/png-resize.ts
@@ -1,6 +1,6 @@
 import { promises as fs } from 'node:fs';
 import { AppError } from '@agent-device/kernel/errors';
-import { PNG } from './png-codec.ts';
+import { PNG } from './png.ts';
 import { decodePngAsync, encodePngAsync } from './png-worker-client.ts';
 
 /**
