+++ src/utils/png-worker-client.ts
@@ -2,8 +2,7 @@ import { Worker } from 'node:worker_threads';
 import { emitDiagnostic } from './diagnostics.ts';
 import { AppError, toAppErrorCode } from '@agent-device/kernel/errors';
 import { resolveInternalEntryModulePath } from './internal-entry.ts';
-import { PNG } from './png-codec.ts';
-import { decodePng } from './png.ts';
+import { decodePng, PNG } from './png.ts';
 import {
   computeScreenshotDiffPixels,
   type ScreenshotDiffPixelsJob,
@@ -201,7 +200,9 @@ export async function decodePngAsync(buffer: Buffer, label: string): Promise<PNG
     const png = decodePng(buffer, label);
     return { kind: 'decode', width: png.width, height: png.height, data: png.data };
   });
-  return new PNG({ width: result.width, height: result.height, data: toBuffer(result.data) });
+  const png = new PNG({ width: result.width, height: result.height });
+  png.data = toBuffer(result.data);
+  return png;
 }
 
 export async function encodePngAsync(png: PNG): Promise<Buffer> {
