+++ src/utils/png-worker.ts
@@ -1,7 +1,6 @@
 import { parentPort } from 'node:worker_threads';
 import { normalizeError } from '@agent-device/kernel/errors';
-import { PNG } from './png-codec.ts';
-import { decodePng } from './png.ts';
+import { decodePng, PNG } from './png.ts';
 import { computeScreenshotDiffPixels } from './screenshot-diff-pixels.ts';
 import { computePngRgbDifference } from './png-rgb-difference.ts';
 import {
@@ -24,11 +23,8 @@ function runJob(request: PngWorkerRequest): PngWorkerJobResult {
       return { kind: 'decode', width: png.width, height: png.height, data: png.data };
     }
     case 'encode': {
-      const png = new PNG({
-        width: request.width,
-        height: request.height,
-        data: toBuffer(request.data),
-      });
+      const png = new PNG({ width: request.width, height: request.height });
+      png.data = toBuffer(request.data);
       return { kind: 'encode', png: PNG.sync.write(png) };
     }
     case 'rgb-difference': {
